Move the testing code from sherpa.optmethods.opt to scripts/opt/
